requirements.txt
pandas
larousse-api-sunbro